COVID Portfolio Project Snippet.sql

+++ COVID Portfolio Project Snippet.sql
@@ -75,7 +75,7 @@ ORDER BY 1,2;
 -- Looking at Total Population Vs Vaccinations
 
 SELECT dea.continent, dea.location,dea.date, dea.population, vac.new_vaccinations
-, SUM(CONVERT(int,vac.new_vaccinations)) OVER (Partition by dea.location Order by dea.location, dea.date) as RollingPeopleVaccinated
+, SUM(CONVERT(bigint,vac.new_vaccinations)) OVER (Partition by dea.location Order by dea.location, dea.date) as RollingPeopleVaccinated
 FROM PortfolioProject..CovidDeaths dea
 Join PortfolioProject..CovidVaccinations vac
 	On dea.location = vac.location
